+++ apps/ensindexer/src/plugins/tokenscope/event-handlers.ts
@@ -1,11 +1,15 @@
 import config from "@/config";
 import { PluginName } from "@ensnode/ensnode-sdk";
 
-import attach_NFTs from "./handlers/NFTs";
+import attach_BaseRegistrars from "./handlers/BaseRegistrars";
+import attach_NameWrapper from "./handlers/NameWrapper";
 import attach_Seaport from "./handlers/Seaport";
+import attach_ThreeDNSToken from "./handlers/ThreeDNSToken";
 
 // conditionally attach event handlers when Ponder executes this file
 if (config.plugins.includes(PluginName.TokenScope)) {
-  attach_NFTs();
+  attach_BaseRegistrars();
+  attach_NameWrapper();
+  attach_ThreeDNSToken();
   attach_Seaport();
 }
